refactor: package depth 추가 - racing / racing.cars 패키지
