+++ src/test/java/step5/domain/ScoreBoardTest.java
@@ -1,12 +1,11 @@
 package step5.domain;
 
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step5.domain.strategy.Movable;
+import step5.domain.racing.Cars;
+import step5.domain.racing.cars.Car;
+import step5.domain.racing.cars.ScoreBoard;
 
 import static org.assertj.core.api.Assertions.assertThat;
